refactor: move ProcessFactory to `wdio-xvfb`
