+++ packages/wdio-local-runner/src/index.ts
@@ -4,13 +4,11 @@ import { xvfb } from '@wdio/xvfb'
 import type { Workers } from '@wdio/types'
 
 import WorkerInstance from './worker.js'
-import { ProcessFactory } from './processFactory.js'
 import { SHUTDOWN_TIMEOUT, BUFFER_OPTIONS } from './constants.js'
 
 const log = logger('@wdio/local-runner')
 
 export type { WorkerInstance }
-export { ProcessFactory }
 
 export interface RunArgs extends Workers.WorkerRunPayload {
     command: string;
